feat(ui): graph save neo4j implementation
